// src/app/model/employee-dashboard.model.ts
export class Employee {
  id?: number = 0;
  firstname: string = '';
  lastname: string = '';
  birthdate: string = '';
  email: string='';
  contact: string='';
  gender: string = '';
  education: string = '';
  department: string = '';
  jobExperience: number = 0;
  salary: number = 0;
  profile: string = '';
  joiningdate: string='';
}